Refactor token and origin logo rendering logic.
